feat: [prd-ui-integration-test-operation-003] - [Split browser-backed integration coverage into feature-oriented files with a shared harness]
